# perf: Add upfront partitioning in `ColumnChunkMetadata`