--- a/crates/polars-io/src/parquet/read/mod.rs
+++ b/crates/polars-io/src/parquet/read/mod.rs
@@ -16,7 +16,6 @@
 
 #[cfg(feature = "cloud")]
 mod async_impl;
-mod metadata;
 mod mmap;
 mod options;
 mod predicates;
@@ -40,7 +39,6 @@ pub use reader::{BatchedParquetReader, ParquetReader};
 pub use utils::materialize_empty_df;
 
 pub mod _internal {
-    pub use super::metadata::PartitionedColumnChunkMD;
     pub use super::mmap::to_deserializer;
     pub use super::predicates::read_this_row_group;
 }