--- a/crates/polars-io/src/parquet/read/predicates.rs
+++ b/crates/polars-io/src/parquet/read/predicates.rs
@@ -1,7 +1,7 @@
 use polars_core::prelude::*;
 use polars_parquet::read::statistics::{deserialize, Statistics};
+use polars_parquet::read::RowGroupMetaData;
 
-use crate::parquet::read::metadata::PartitionedColumnChunkMD;
 use crate::predicates::{BatchStats, ColumnStats, PhysicalIoExpr};
 
 impl ColumnStats {
@@ -17,18 +17,20 @@ impl ColumnStats {
 
 /// Collect the statistics in a row-group
 pub(crate) fn collect_statistics(
-    part_md: &PartitionedColumnChunkMD,
+    md: &RowGroupMetaData,
     schema: &ArrowSchema,
 ) -> PolarsResult<Option<BatchStats>> {
     // TODO! fix this performance. This is a full sequential scan.
     let stats = schema
         .iter_values()
-        .map(|field| match part_md.get_partitions(&field.name) {
-            Some(md) => {
-                let st = deserialize(field, &md)?;
-                Ok(ColumnStats::from_arrow_stats(st, field))
-            },
-            None => Ok(ColumnStats::new(field.into(), None, None, None)),
+        .map(|field| {
+            let iter = md.columns_under_root_iter(&field.name);
+
+            Ok(if iter.len() == 0 {
+                ColumnStats::new(field.into(), None, None, None)
+            } else {
+                ColumnStats::from_arrow_stats(deserialize(field, iter)?, field)
+            })
         })
         .collect::<PolarsResult<Vec<_>>>()?;
 
@@ -39,18 +41,18 @@ pub(crate) fn collect_statistics(
     Ok(Some(BatchStats::new(
         Arc::new(Schema::from_arrow_schema(schema)),
         stats,
-        Some(part_md.num_rows()),
+        Some(md.num_rows()),
     )))
 }
 
 pub fn read_this_row_group(
     predicate: Option<&dyn PhysicalIoExpr>,
-    part_md: &PartitionedColumnChunkMD,
+    md: &RowGroupMetaData,
     schema: &ArrowSchema,
 ) -> PolarsResult<bool> {
     if let Some(pred) = predicate {
         if let Some(pred) = pred.as_stats_evaluator() {
-            if let Some(stats) = collect_statistics(part_md, schema)? {
+            if let Some(stats) = collect_statistics(md, schema)? {
                 let should_read = pred.should_read(&stats);
                 // a parquet file may not have statistics of all columns
                 if matches!(should_read, Ok(false)) {