--- a/crates/polars-parquet/src/parquet/metadata/row_metadata.rs
+++ b/crates/polars-parquet/src/parquet/metadata/row_metadata.rs
@@ -1,38 +1,64 @@
+use hashbrown::hash_map::RawEntryMut;
 use parquet_format_safe::RowGroup;
-#[cfg(feature = "serde_types")]
-use serde::{Deserialize, Serialize};
+use polars_utils::aliases::{InitHashMaps, PlHashMap};
+use polars_utils::idx_vec::UnitVec;
+use polars_utils::pl_str::PlSmallStr;
+use polars_utils::unitvec;
 
-use super::column_chunk_metadata::ColumnChunkMetaData;
+use super::column_chunk_metadata::{column_metadata_byte_range, ColumnChunkMetadata};
 use super::schema_descriptor::SchemaDescriptor;
 use crate::parquet::error::{ParquetError, ParquetResult};
 use crate::parquet::write::ColumnOffsetsMetadata;
 
+type ColumnLookup = PlHashMap<PlSmallStr, UnitVec<usize>>;
+
+trait InitColumnLookup {
+    fn add_column(&mut self, index: usize, column: &ColumnChunkMetadata);
+}
+
+impl InitColumnLookup for ColumnLookup {
+    #[inline(always)]
+    fn add_column(&mut self, index: usize, column: &ColumnChunkMetadata) {
+        let root_name = &column.descriptor().path_in_schema[0];
+
+        match self.raw_entry_mut().from_key(root_name) {
+            RawEntryMut::Vacant(slot) => {
+                slot.insert(root_name.clone(), unitvec![index]);
+            },
+            RawEntryMut::Occupied(mut slot) => {
+                slot.get_mut().push(index);
+            },
+        };
+    }
+}
+
 /// Metadata for a row group.
 #[derive(Debug, Clone, Default)]
-#[cfg_attr(feature = "serde_types", derive(Deserialize, Serialize))]
+// #[cfg_attr(feature = "serde_types", derive(serde::Deserialize, serde::Serialize))]
 pub struct RowGroupMetaData {
-    columns: Vec<ColumnChunkMetaData>,
+    columns: Vec<ColumnChunkMetadata>,
+    column_lookup: PlHashMap<PlSmallStr, UnitVec<usize>>,
     num_rows: usize,
     total_byte_size: usize,
+    full_byte_range: core::ops::Range<u64>,
 }
 
 impl RowGroupMetaData {
-    /// Create a new [`RowGroupMetaData`]
-    pub fn new(
-        columns: Vec<ColumnChunkMetaData>,
-        num_rows: usize,
-        total_byte_size: usize,
-    ) -> RowGroupMetaData {
-        Self {
-            columns,
-            num_rows,
-            total_byte_size,
-        }
+    #[inline(always)]
+    pub fn n_columns(&self) -> usize {
+        self.columns.len()
     }
 
-    /// Returns slice of column chunk metadata.
-    pub fn columns(&self) -> &[ColumnChunkMetaData] {
-        &self.columns
+    /// Fetch all columns under this root name.
+    pub fn columns_under_root_iter(
+        &self,
+        root_name: &str,
+    ) -> impl ExactSizeIterator<Item = &ColumnChunkMetadata> + DoubleEndedIterator {
+        self.column_lookup
+            .get(root_name)
+            .unwrap()
+            .iter()
+            .map(|&x| &self.columns[x])
     }
 
     /// Number of rows in this row group.
@@ -53,6 +79,14 @@ impl RowGroupMetaData {
             .sum::<usize>()
     }
 
+    pub fn full_byte_range(&self) -> core::ops::Range<u64> {
+        self.full_byte_range.clone()
+    }
+
+    pub fn byte_ranges_iter(&self) -> impl '_ + ExactSizeIterator<Item = core::ops::Range<u64>> {
+        self.columns.iter().map(|x| x.byte_range())
+    }
+
     /// Method to convert from Thrift.
     pub(crate) fn try_from_thrift(
         schema_descr: &SchemaDescriptor,
@@ -63,19 +97,42 @@ impl RowGroupMetaData {
         }
         let total_byte_size = rg.total_byte_size.try_into()?;
         let num_rows = rg.num_rows.try_into()?;
+
+        let mut column_lookup = ColumnLookup::with_capacity(rg.columns.len());
+        let mut full_byte_range = if let Some(first_column_chunk) = rg.columns.first() {
+            let Some(metadata) = &first_column_chunk.meta_data else {
+                return Err(ParquetError::oos("Column chunk requires metadata"));
+            };
+            column_metadata_byte_range(metadata)
+        } else {
+            0..0
+        };
+
         let columns = rg
             .columns
             .into_iter()
             .zip(schema_descr.columns())
-            .map(|(column_chunk, descriptor)| {
-                ColumnChunkMetaData::try_from_thrift(descriptor.clone(), column_chunk)
+            .enumerate()
+            .map(|(i, (column_chunk, descriptor))| {
+                let column =
+                    ColumnChunkMetadata::try_from_thrift(descriptor.clone(), column_chunk)?;
+
+                column_lookup.add_column(i, &column);
+
+                let byte_range = column.byte_range();
+                full_byte_range = full_byte_range.start.min(byte_range.start)
+                    ..full_byte_range.end.max(byte_range.end);
+
+                Ok(column)
             })
             .collect::<ParquetResult<Vec<_>>>()?;
 
         Ok(RowGroupMetaData {
             columns,
+            column_lookup,
             num_rows,
             total_byte_size,
+            full_byte_range,
         })
     }