… Cargo.lock
… crates/polars-io/src/parquet/read/metadata.rs
… crates/polars-parquet/Cargo.toml
… crates/polars-parquet/src/parquet/read/column/mod.rs
… crates/polars-parquet/src/parquet/read/mod.rs
… crates/polars-parquet/src/parquet/read/page/reader.rs
… crates/polars-parquet/src/parquet/read/page/stream.rs
… crates/polars-stream/src/nodes/parquet_source.rs
… crates/polars/tests/it/io/parquet/read/mod.rs
… crates/polars/tests/it/io/parquet/read/row_group.rs
… crates/polars/tests/it/io/parquet/write/mod.rs