@ritchie46 2024-09-07
Wow, that's an impressive improvement. It shows we had a lot of accidental linear scans allover. :)